RPGT-8: Extract config and database pool into dedicated modules
